add use-dev-tty feature flag
