+++ Cargo.toml
@@ -29,6 +29,7 @@ all-features = true
 default = ["bracketed-paste"]
 bracketed-paste = []
 event-stream = ["futures-core"]
+use-dev-tty = ["filedescriptor"]
 
 #
 # Shared dependencies
@@ -57,7 +58,9 @@ crossterm_winapi = "0.9"
 [target.'cfg(unix)'.dependencies]
 libc = "0.2"
 signal-hook = { version = "0.3.13" }
-filedescriptor = "0.8"
+filedescriptor = { version = "0.8", optional = true }
+mio = { version = "0.8", features = ["os-poll"] }
+signal-hook-mio = { version = "0.2.3", features = ["support-v0_8"] }
 
 #
 # Dev dependencies (examples, ...)
