+++ src/event/sys/unix/waker.rs
@@ -1,38 +1,11 @@
-use std::{
-    io::Write,
-    os::unix::net::UnixStream,
-    sync::{Arc, Mutex},
-};
+#[cfg(feature = "use-dev-tty")]
+pub(crate) mod tty;
 
-use crate::Result;
+#[cfg(not(feature = "use-dev-tty"))]
+pub(crate) mod mio;
 
-/// Allows to wake up the EventSource::try_read() method.
-#[derive(Clone, Debug)]
-pub(crate) struct Waker {
-    inner: Arc<Mutex<UnixStream>>,
-}
+#[cfg(feature = "use-dev-tty")]
+pub(crate) use self::tty::Waker;
 
-impl Waker {
-    /// Create a new `Waker`.
-    pub(crate) fn new(writer: UnixStream) -> Self {
-        Self {
-            inner: Arc::new(Mutex::new(writer)),
-        }
-    }
-
-    /// Wake up the [`Poll`] associated with this `Waker`.
-    ///
-    /// Readiness is set to `Ready::readable()`.
-    pub(crate) fn wake(&self) -> Result<()> {
-        self.inner.lock().unwrap().write(&[0])?;
-        Ok(())
-    }
-
-    /// Resets the state so the same waker can be reused.
-    ///
-    /// This function is not impl
-    #[allow(dead_code, clippy::clippy::unnecessary_wraps)]
-    pub(crate) fn reset(&self) -> Result<()> {
-        Ok(())
-    }
-}
+#[cfg(not(feature = "use-dev-tty"))]
+pub(crate) use self::mio::Waker;
